separted testclasses for ime actions and error tests, completed testcase testActionKey()
